refactor: separate action by table
